favorites update

+++ src/Controllers/FavoritesDessertsController.js
@@ -11,7 +11,8 @@ class favoritesDessertsController {
   
   async delete(request, response) {
     const { id } = request.params
-    await knex("favoritesDesserts").where({ dessert_id: id }).delete()
+    const user_id = request.user.id
+    await knex("favoritesDesserts").where({ dessert_id: id }).where({ user_id }).delete()
     return response.json()
   }
   async show(request, response) {
